move user verification to endpoint from trigger
